+++ foolbox/batch_attacks/iterative_projected_gradient.py
@@ -5,7 +5,7 @@ import logging
 import warnings
 
 from ..attacks.base import Attack
-from ..attacks.base import generator_call_decorator
+from .decorator import generator_call_decorator
 from .. import distances
 from ..utils import crossentropy
 from .. import nprng
